+++ src/bson/bson-utf8.h
@@ -40,7 +40,7 @@ bson_unichar_t  bson_utf8_get_char        (const char     *utf8);
 const char     *bson_utf8_next_char       (const char     *utf8);
 void            bson_utf8_from_unichar    (bson_unichar_t  unichar,
                                            char            utf8[6],
-                                           uint32_t       *len);
+                                           size_t         *len);
 
 
 BSON_END_DECLS
